requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
audioread==2.1.8
autocorrect==0.3.0
blitzdb==0.4.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
checkmate==0.2.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
eyeD3==0.8.0a5
ffmpeg==1.4
gast==0.2.2
google-pasta==0.1.7
grako==3.99.9
grpcio==1.22.0
h5py==2.9.0
idna==2.8
isort==4.3.21
joblib==0.13.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libmagic==1.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
matplotlib==3.0.0
mccabe==0.6.1
nltk==3.4.3
numba==0.45.1
numpy==1.17.0
pandas==0.25.0
pep8==1.7.1
protobuf==3.9.0
pyAudioAnalysis==0.2.5
pycparser==2.19
pydub==0.23.1
pyflakes==2.1.1
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.1
requests==2.22.0
resampy==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
simplejson==3.16.0
six==1.12.0
SoundFile==0.10.2
speechpy==2.4
SQLAlchemy==1.3.8
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.32.1
typed-ast==1.4.0
urllib3==1.25.3
validators==0.13.0
Werkzeug==0.15.5
wrapt==1.11.2
